Other fixes related to 32-bin signed integers

+++ yrt-pet/src/operators/OperatorPsf.cpp
@@ -293,10 +293,10 @@ void OperatorPsf::convolve(const Image* in, Image* out,
 	const ImageParams& params = in->getParams();
 	ASSERT_MSG(params.isSameDimensionsAs(out->getParams()),
 	           "Dimensions mismatch between the two images");
-	const int nx = params.nx;
-	const int ny = params.ny;
-	const int nz = params.nz;
-	const int nt = params.nt;
+	const ssize_t nx = params.nx;
+	const ssize_t ny = params.ny;
+	const ssize_t nz = params.nz;
+	const ssize_t nt = params.nt;
 
 	const size_t frameSize = static_cast<size_t>(nx) * ny * nz;
 	const size_t sizeBuffer = std::max(std::max(nx, ny), nz);
@@ -309,26 +309,26 @@ void OperatorPsf::convolve(const Image* in, Image* out,
 	const int kerIndexCenteredY = kerSize[1] / 2;
 	const int kerIndexCenteredZ = kerSize[2] / 2;
 
-	for (int t = 0; t < nt; t++)
+	for (ssize_t t = 0; t < nt; t++)
 	{
 		const float* inPtr = in->getRawPointer() + t * frameSize;
 		float* outPtr = out->getRawPointer() + t * frameSize;
 
-		for (int k = 0; k < nz; k++)
+		for (ssize_t k = 0; k < nz; k++)
 		{
-			for (int j = 0; j < ny; j++)
+			for (ssize_t j = 0; j < ny; j++)
 			{
-				for (int i = 0; i < nx; i++)
+				for (ssize_t i = 0; i < nx; i++)
 				{
 					m_buffer_tmp[i] = inPtr[IDX3(i, j, k, nx, ny)];
 				}
-				for (int i = 0; i < nx; i++)
+				for (ssize_t i = 0; i < nx; i++)
 				{
 					float sum = 0.0f;
 					for (int kk = -kerIndexCenteredX; kk <= kerIndexCenteredX;
 					     kk++)
 					{
-						const int r = util::circular(nx, i - kk);
+						const ssize_t r = util::circular(nx, i - kk);
 						sum +=
 						    kernelX[kk + kerIndexCenteredX] * m_buffer_tmp[r];
 					}
@@ -337,21 +337,21 @@ void OperatorPsf::convolve(const Image* in, Image* out,
 			}
 		}
 
-		for (int k = 0; k < nz; k++)
+		for (ssize_t k = 0; k < nz; k++)
 		{
-			for (int i = 0; i < nx; i++)
+			for (ssize_t i = 0; i < nx; i++)
 			{
-				for (int j = 0; j < ny; j++)
+				for (ssize_t j = 0; j < ny; j++)
 				{
 					m_buffer_tmp[j] = outPtr[IDX3(i, j, k, nx, ny)];
 				}
-				for (int j = 0; j < ny; j++)
+				for (ssize_t j = 0; j < ny; j++)
 				{
 					float sum = 0.0f;
 					for (int kk = -kerIndexCenteredY; kk <= kerIndexCenteredY;
 					     kk++)
 					{
-						const int r = util::circular(ny, j - kk);
+						const ssize_t r = util::circular(ny, j - kk);
 						sum +=
 						    kernelY[kk + kerIndexCenteredY] * m_buffer_tmp[r];
 					}
@@ -360,21 +360,21 @@ void OperatorPsf::convolve(const Image* in, Image* out,
 			}
 		}
 
-		for (int i = 0; i < nx; i++)
+		for (ssize_t i = 0; i < nx; i++)
 		{
-			for (int j = 0; j < ny; j++)
+			for (ssize_t j = 0; j < ny; j++)
 			{
-				for (int k = 0; k < nz; k++)
+				for (ssize_t k = 0; k < nz; k++)
 				{
 					m_buffer_tmp[k] = outPtr[IDX3(i, j, k, nx, ny)];
 				}
-				for (int k = 0; k < nz; k++)
+				for (ssize_t k = 0; k < nz; k++)
 				{
 					float sum = 0.0f;
 					for (int kk = -kerIndexCenteredZ; kk <= kerIndexCenteredZ;
 					     kk++)
 					{
-						const int r = util::circular(nz, k - kk);
+						const ssize_t r = util::circular(nz, k - kk);
 						sum +=
 						    kernelZ[kk + kerIndexCenteredZ] * m_buffer_tmp[r];
 					}
